third_party/WebKit/LayoutTests/inspector/console/console-format-perfomance-expected.txt

CONSOLE MESSAGE: line 7: [object PerformanceTiming]
CONSOLE MESSAGE: line 8: [object MemoryInfo]
CONSOLE MESSAGE: line 7: [object PerformanceTiming]
CONSOLE MESSAGE: line 8: [object MemoryInfo]
Tests that console nicely formats perfomance getters.

console-format-perfomance.html:7 PerformanceTiming
console-format-perfomance.html:8 MemoryInfo
console-format-perfomance.html:7 PerformanceTiming {navigationStart: <number>, unloadEventStart: <number>, unloadEventEnd: <number>, redirectStart: <number>, redirectEnd: <number>, …}
console-format-perfomance.html:8 MemoryInfo {totalJSHeapSize: <number>, usedJSHeapSize: <number>, jsHeapSizeLimit: <number>}
